+++ app/static/lib/main.js
@@ -1,24 +1,23 @@
-export const version = "0.0.2";
-export const versionDate = "20 February 2025";
+export const version = "0.1.0";
+export const versionDate = "5 April 2025";
 
 import NS from "./namespaceManager.js";
-import LDObj from "./LDObj.js";
 import Traverser from "./traverser.js";
 import Graph from "./graph.js";
-import { requestAsJsonLd, fetchTextData } from "./http.js";
+import { fetchTextData } from "./http.js";
 import { vrvOptions } from "./defaults.js";
 
 let objUrl = null;
 
 function traversalsComplete() {
-  console.log("Handling traversals completed");
+  console.log("Handling traversals completed, registry: ", Graph.registry);
   // render the graph using mermaid
-  if (objUrl in Graph.registry) {
+  if (objUrl.href in Graph.registry) {
     let visgraph =
       "graph TD; " + Graph.visualise(Graph.registry[objUrl].expanded, true);
     // remove protocols from the URLs to not upset mermaid
     visgraph = visgraph.replace(/https?:\/\//g, "");
-    Object.keys(Graph.registry).forEach((key) => {
+    for (const key of Object.keys(Graph.registry)) {
       // only replace the first protocol (acting as label), not the second (acting as URL)
       visgraph +=
         "\n click " +
@@ -26,7 +25,7 @@ function traversalsComplete() {
         ' "http://localhost:5001/?obj=' +
         key +
         '" "Tooltip: Foo";';
-    });
+    }
     let jsonDisplay = document.querySelector("#json-display");
     jsonDisplay.textContent = JSON.stringify(
       Graph.registry[objUrl].compacted,
@@ -42,7 +41,9 @@ function traversalsComplete() {
   } else {
     console.error(
       "Object not found in registry while trying to visualise: ",
-      objUrl
+      objUrl.href,
+      Object.keys(Graph.registry),
+      Graph.registry
     );
   }
   // render the MEI using Verovio:
@@ -51,12 +52,12 @@ function traversalsComplete() {
   // and use Verovio to render the respective MEI to SVG inside the div
   let meiTargets = Graph.getMEITargets();
   console.log("Main got MEI targets: ", meiTargets);
-  Object.keys(meiTargets).forEach((target) => {
+  for (const target of Object.keys(meiTargets)) {
     // use fetch to retrieve the target
     fetchTextData(target).then((data) =>
       renderMEI(target, meiTargets[target], data)
     );
-  });
+  }
 
   // render the textual bodies
   let textBodies = Graph.getTextualBodies();
@@ -69,13 +70,13 @@ function traversalsComplete() {
 
 function renderTextualBodies(textBodies) {
   const textDiv = document.getElementById("text-content");
-  textBodies.forEach((b) => {
-    let text = b["@value"];
+  for (const b of Object.keys(textBodies)) {
+    let text = textBodies[b]["@value"];
     let textDivChild = document.createElement("div");
     textDivChild.className = "textual-body";
     textDivChild.innerHTML = text;
     textDiv.appendChild(textDivChild);
-  });
+  }
 }
 
 function renderMEI(uri, fragmentSet, meiData) {
@@ -102,16 +103,24 @@ function renderMEI(uri, fragmentSet, meiData) {
   let svg = vrvToolkit.renderToSVG(pageNum);
   meiDiv.innerHTML += svg;
   // add "highlight" class to each element with the fragment ID
-  fragments.forEach((f) => {
+  for (const f of fragments) {
     let el = meiDiv.querySelector(`[id="${f}"]`);
     if (el) {
       el.classList.add("highlight");
     }
     console.log("element: ", el);
-  });
+  }
 }
 
 document.addEventListener("DOMContentLoaded", async function () {
+  // fill in version and date
+  let versionSpan = document.querySelector("#version");
+  let dateSpan = document.querySelector("#date");
+  if (versionSpan && dateSpan) {
+    versionSpan.innerHTML = version;
+    dateSpan.innerHTML = versionDate;
+  }
+
   Graph.init(NS);
   Traverser.init();
   // add event listener to trigger when traversals are complete
@@ -131,110 +140,3 @@ document.addEventListener("DOMContentLoaded", async function () {
     return;
   }
 });
-
-// event listener to trigger when DOM loaded
-//document.addEventListener("DOMContentLoaded", async function () {
-//  // check for presence of ?oa parameter in URL
-//  const urlParams = new URLSearchParams(window.location.search);
-//  const oaParam = urlParams.get("oa");
-//  const maoParam = urlParams.get("mao");
-//  // if ?oa parameter is present, ensure it is a valid URL
-//  if (oaParam || maoParam) {
-//    const paramType = oaParam ? "oa" : "mao";
-//    const paramValue = paramType === "oa" ? oaParam : maoParam;
-//    const objUrl = new URL(paramValue);
-//    // if valid, request the URL as json-ld
-//    if (objUrl) {
-//      try {
-//        let data = await requestAsJsonLd(objUrl);
-//        console.log("Retrieved JSON data: ", data);
-//        let ldo = new LDObj(data);
-//        ldo
-//          .prepare()
-//          .then(() => {
-//            if (ldo.expanded) {
-//              let graph = V.drawGraph(ldo.expanded);
-//              mermaid.render("fograph", graph).then((svg) => {
-//                document.getElementById("graph").innerHTML = svg.svg;
-//              });
-//            } else {
-//              console.error("Failed to initialize OA");
-//            }
-//            if (ldo.hasTarget()) {
-//              // for each target with type "MEI",
-//              // create a div inside #music-scores
-//              // and use Verovio to render the MEI to SVG inside the div
-//              let meiTargets = Object.keys(ldo.targets).filter((key) => {
-//                return ldo.targets[key].type === "MEI";
-//              });
-//              meiTargets.forEach((target) => {
-//                // use fetch to retrieve the target
-//                fetch(target)
-//                  .then((response) => {
-//                    if (!response.ok) {
-//                      throw new Error(
-//                        "Network response was not ok",
-//                        response,
-//                        target
-//                      );
-//                    }
-//                    return response.text();
-//                  })
-//                  .then((meiData) => {
-//                    // create a new div for the MEI
-//                    let meiDiv = document.createElement("div");
-//                    meiDiv.className = "mei";
-//                    meiDiv.id = target;
-//                    document.getElementById("music-scores").appendChild(meiDiv);
-//                    // use Verovio to render the MEI to SVG inside the div
-//                    let vrvToolkit = new verovio.toolkit();
-//                    vrvToolkit.setOptions(vrvOptions);
-//                    vrvToolkit.loadData(meiData);
-//                    let fragments = Array.from(ldo.targets[target].fragments);
-//                    let meifriendDiv = document.createElement("div");
-//                    meifriendDiv.className = "meifriendLink";
-//                    let link = "https://mei-friend.mdw.ac.at/?file=" + target;
-//                    link += "&select=" + fragments.join(",");
-//                    meifriendDiv.innerHTML = `<a href="${link} target="_blank">Open in mei-friend</a>`;
-//                    meiDiv.innerHTML = "";
-//                    meiDiv.appendChild(meifriendDiv);
-//                    let fragment = fragments[0];
-//                    let pageNum = vrvToolkit.getPageWithElement(fragment);
-//                    let svg = vrvToolkit.renderToSVG(pageNum);
-//                    meiDiv.innerHTML += svg;
-//                    // add "highlight" class to each element with the fragment ID
-//                    fragments.forEach((f) => {
-//                      let el = meiDiv.querySelector(`[id="${f}"]`);
-//                      if (el) {
-//                        el.classList.add("highlight");
-//                      }
-//                      console.log("element: ", el);
-//                    });
-//                  })
-//                  .catch((error) => {
-//                    console.error("Failed to load MEI data: ", error);
-//                  });
-//              });
-//            }
-//            if (ldo.hasTextualBody()) {
-//              const textDiv = document.getElementById("text-content");
-//              ldo.getTextualBodies().forEach((b) => {
-//                let text = b["@value"];
-//                let textDivChild = document.createElement("div");
-//                textDivChild.className = "textual-body";
-//                textDivChild.innerHTML = text;
-//                textDiv.appendChild(textDivChild);
-//              });
-//            }
-//          })
-//          .catch((error) => {
-//            console.error("Failed to prepare OA:", error);
-//          });
-//      } catch (error) {
-//        console.error("Failed to retrieve OA as JSON-LD: ", error);
-//      }
-//    } else {
-//      console.error("Invalid URL");
-//    }
-//  }
-// })