+++ app/static/lib/traverser.js
@@ -18,16 +18,21 @@ export default class Traverser {
   static finishedOne() {
     // mark that one traversal round, i.e. a full fetchAndRegister cycle, is finished
     this.numTraversals--;
+  }
+
+  static checkFinished() {
+    // check if all traversals are finished
     if (this.numTraversals <= 0) {
       // emit event to indicate all traversals are complete
       console.log("All traversals complete");
       document.dispatchEvent(this.completionEvent);
     }
   }
+
   static async fetchAndRegister(urls = []) {
     console.log("fetchAndRegister called with urls: ", urls);
-    console.log("numTraversals before: ", this.numTraversals);
     for (const url of urls) {
+      console.log("numTraversals before: ", this.numTraversals);
       // fetch the object from the URL
       // and register it in the registry
       this.numTraversals = this.numTraversals < 0 ? 1 : this.numTraversals + 1;
@@ -40,11 +45,9 @@ export default class Traverser {
       );
       const data = await requestAsJsonLd(url);
       const obj = new LDObj(data);
-      console.log("Fetched object: ", obj);
-      console.log("Based on data ", data);
       await obj.prepare();
       Graph.register(obj, url);
-      console.log("Traversals remaining: ", this.numTraversals);
+      this.checkFinished();
     }
   }
 }