+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # SSV-Viewer (to be renamed) CHANGELOG
 
+### 0.1.0 - 5 Apr 2025
+* Visualisation of MAO as well as OA
+* Add navigation to visualisation
+* Add JSON listing
+* Add Verovio score with target highlight
+
 ### 0.0.2 - 20 Feb 2025
 * Ugly but functional visualisation of OA using Mermaid.js